essai

+++ APP_FILMS_164/films/gestion_films_crud.py
@@ -61,13 +61,13 @@ def films_afficher(id_film_sel):
                     # Si l'utilisateur change l'id_film dans l'URL et qu'il ne correspond à aucun film
                     flash(f"Le film {id_film_sel} demandé n'existe pas !!", "warning")
                 else:
-                    flash(f"Données films et genres affichés !!", "success")
+                    flash(f"Données sur les fournisseurs affichés", "success")
 
         except Exception as Exception_films_genres_afficher:
             raise ExceptionFilmsGenresAfficher(f"fichier : {Path(__file__).name}  ;  {films_afficher.__name__} ;"
                                                f"{Exception_films_genres_afficher}")
 
-    print("films_genres_afficher  ", data_genres_films_afficher)
+    print("films_afficher  ", data_genres_films_afficher)
     # Envoie la page "HTML" au serveur.
     return render_template("films/films_afficher.html", data=data_genres_films_afficher)
 
@@ -84,8 +84,6 @@ Remarque :  Dans le champ "nom_film_update_wtf" du formulaire "films/films_updat
             le contrôle de la saisie s'effectue ici en Python dans le fichier ""
             On ne doit pas accepter un champ vide.
 """
-
-
 @app.route("/film_add", methods=['GET', 'POST'])
 def film_add_wtf():
     form_add_film = FormWTFAddFilm()
@@ -99,7 +97,7 @@ def film_add_wtf():
                 valeurs_insertion_dictionnaire = {
                     "value_nom_film": nom_fournisseur_add_wtf,
                     "value_adresse": adresse_fournisseur_add_wtf,
-                    "Value_telephone": numero_telephone_fournisseur_add_wtf
+                    "Value_telephone": numero_telephone_fournisseur_add_wtf,
                 }
                 print("valeurs_insertion_dictionnaire ", valeurs_insertion_dictionnaire)
 
@@ -110,12 +108,13 @@ def film_add_wtf():
                 flash("Données insérées !!", "success")
                 print("Données insérées !!")
 
-                return redirect(url_for('fournisseur_produit_afficher'))
+                # Redirection vers le bon endpoint avec le bon paramètre
+                return redirect(url_for('films_afficher', id_film_sel=0))
 
-        except Exception as Exception_genres_ajouter_wtf:
+        except Exception as Exception_fournisseur_ajouter_wtf:
             raise ExceptionGenresAjouterWtf(f"fichier : {Path(__file__).name}  ;  "
                                             f"{film_add_wtf.__name__} ; "
-                                            f"{Exception_genres_ajouter_wtf}")
+                                            f"{Exception_fournisseur_ajouter_wtf}")
 
     return render_template("films/film_add_wtf.html", form_add_film=form_add_film)
 
@@ -136,48 +135,46 @@ Remarque :  Dans le champ "nom_film_update_wtf" du formulaire "films/films_updat
 
 @app.route("/film_update", methods=['GET', 'POST'])
 def film_update_wtf():
-    id_film_update = request.values['id_film_btn_edit_html']
+    id_film_update = request.values.get('id_film_btn_edit_html')  # Utilisation de get() pour éviter les erreurs KeyError
     form_update_film = FormWTFUpdateFilm()
     try:
         if request.method == "POST" and form_update_film.submit.data:
-            nom_fournissuer_update = form_update_film.nom_fournissuer_update.data
+            nom_fournisseur_update = form_update_film.nom_fournisseur_update.data
             adresse_fournisseur_update = form_update_film.adresse_fournisseur_update.data
             numero_telephone_fournisseur_update = form_update_film.numero_telephone_fournisseur_update.data
 
-
             valeur_update_dictionnaire = {
                 "value_id_film": id_film_update,
-                "value_nom_film": nom_fournissuer_update,
+                "value_nom_film": nom_fournisseur_update,
                 "value_adresse_fournisseur": adresse_fournisseur_update,
                 "value_numero_telephone_fournisseur": numero_telephone_fournisseur_update,
             }
             print("valeur_update_dictionnaire ", valeur_update_dictionnaire)
 
             str_sql_update_nom_film = """UPDATE t_fournisseur SET nom_fournisseur = %(value_nom_film)s,
-                                                        adresse_fournisseur = %(value_adresse_fournisseur)s,
-                                                        numero_telephone_fournisseur = %(value_numero_telephone_fournisseur)s,
-                                                        WHERE ID_fournisseur = %(value_id_film)s"""
+                                                            adresse_fournisseur = %(value_adresse_fournisseur)s,
+                                                            numero_telephone_fournisseur = %(value_numero_telephone_fournisseur)s
+                                                            WHERE ID_fournisseur = %(value_id_film)s"""
             with DBconnection() as mconn_bd:
                 mconn_bd.execute(str_sql_update_nom_film, valeur_update_dictionnaire)
 
             flash("Donnée mise à jour !!", "success")
             print("Donnée mise à jour !!")
 
-            return redirect(url_for('fournisseur_produit_afficher'))
+            return redirect(url_for('films_afficher', id_film_sel=0))
         elif request.method == "GET":
-            str_sql_id_film = "SELECT * FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"
-            valeur_select_dictionnaire = {"value_id_film": id_film_update}
-            with DBconnection() as mybd_conn:
-                mybd_conn.execute(str_sql_id_film, valeur_select_dictionnaire)
-            data_film = mybd_conn.fetchone()
-            print("data_film ", data_film, " type ", type(data_film), " genre ",
-                  data_film["nom_film"])
+            if id_film_update:  # Vérification si l'ID du film est présent
+                str_sql_id_film = "SELECT * FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"
+                valeur_select_dictionnaire = {"value_id_film": id_film_update}
+                with DBconnection() as mybd_conn:
+                    mybd_conn.execute(str_sql_id_film, valeur_select_dictionnaire)
+                    data_film = mybd_conn.fetchone()
+                    print("data_film ", data_film, " type ", type(data_film))
 
-            form_update_film.nom_fournissuer_update.data = data_film["nom_film"]
-            form_update_film.adresse_fournisseur_update.data = data_film["duree_film"]
-            print(f" duree film  ", data_film["duree_film"], "  type ", type(data_film["duree_film"]))
-            form_update_film.numero_telephone_fournisseur_update.data = data_film["description_film"]
-
+                    # Assurez-vous que les noms des champs dans le formulaire correspondent exactement aux attributs du formulaire
+                    form_update_film.nom_fournisseur_update.data = data_film["nom_fournisseur"]
+                    form_update_film.adresse_fournisseur_update.data = data_film["adresse_fournisseur"]
+                    form_update_film.numero_telephone_fournisseur_update.data = data_film["numero_telephone_fournisseur"]
 
     except Exception as Exception_film_update_wtf:
         raise ExceptionFilmUpdateWtf(f"fichier : {Path(__file__).name}  ;  "
@@ -185,7 +182,6 @@ def film_update_wtf():
                                      f"{Exception_film_update_wtf}")
 
     return render_template("films/film_update_wtf.html", form_update_film=form_update_film)
-
 
 """Effacer(delete) un film qui a été sélectionné dans le formulaire "fournisseur_produit_afficher.html"
 Auteur : OM 2022.04.11
@@ -204,53 +200,64 @@ Remarque :  Dans le champ "nom_film_delete_wtf" du formulaire "films/film_delete
 def film_delete_wtf():
     data_film_delete = None
     btn_submit_del = None
-    id_film_delete = request.values['id_film_btn_delete_html']
+    id_film_delete = request.values.get('id_film_btn_delete_html')
     form_delete_film = FormWTFDeleteFilm()
+
     try:
         if form_delete_film.submit_btn_annuler.data:
-            return redirect(url_for("films_genres_afficher"))
+            return redirect(url_for("films_afficher", id_film_sel=0))
 
         if form_delete_film.submit_btn_conf_del_film.data:
-            data_film_delete = session['data_film_delete']
-            print("data_film_delete ", data_film_delete)
-
-            flash("Effacer le film de façon définitive de la BD !!!", "danger")
+            # Affiche un message Flash pour confirmer la suppression
+            flash("Voulez-vous vraiment supprimer ce film ?", "warning")
             btn_submit_del = True
 
         if form_delete_film.submit_btn_del_film.data:
-            valeur_delete_dictionnaire = {"value_id_film": id_film_delete}
-            print("valeur_delete_dictionnaire ", valeur_delete_dictionnaire)
+            # Vérifie si l'ID du film à supprimer est disponible
+            if id_film_delete:
+                valeur_delete_dictionnaire = {"value_id_film": id_film_delete}
 
-            str_sql_delete_fk_film_genre = """DELETE FROM t_fournisseur_produit WHERE ID_fournisseur = %(value_id_film)s"""
-            str_sql_delete_film = """DELETE FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"""
-            with DBconnection() as mconn_bd:
-                mconn_bd.execute(str_sql_delete_fk_film_genre, valeur_delete_dictionnaire)
-                mconn_bd.execute(str_sql_delete_film, valeur_delete_dictionnaire)
+                # Exécute la requête SQL pour supprimer le film de la base de données
+                str_sql_delete_film = """DELETE FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"""
+                with DBconnection() as mconn_bd:
+                    mconn_bd.execute(str_sql_delete_film, valeur_delete_dictionnaire)
 
-            flash("Film définitivement effacé !!", "success")
-            print("Film définitivement effacé !!")
+                # Affiche un message Flash pour indiquer que le film a été supprimé avec succès
+                flash("Film définitivement effacé !!", "success")
+                print("Film définitivement effacé !!")
 
-            return redirect(url_for('films_genres_afficher'))
+                # Redirige vers la page affichant tous les films
+                return redirect(url_for('films_afficher', id_film_sel=0))
+            else:
+                # Si l'ID du film à supprimer est manquant, affiche un message d'avertissement
+                flash("Impossible de trouver l'ID du film à supprimer.", "warning")
+
         if request.method == "GET":
-            valeur_select_dictionnaire = {"value_id_film": id_film_delete}
-            print(id_film_delete, type(id_film_delete))
+            if id_film_delete:
+                # Récupère les données du film à supprimer
+                valeur_select_dictionnaire = {"value_id_film": id_film_delete}
+                str_sql_genres_films_delete = """SELECT * FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"""
+                with DBconnection() as mydb_conn:
+                    mydb_conn.execute(str_sql_genres_films_delete, valeur_select_dictionnaire)
+                    data_film_delete = mydb_conn.fetchall()
+                    print("data_film_delete...", data_film_delete)
+                    session['data_film_delete'] = data_film_delete
 
-            str_sql_genres_films_delete = """SELECT * FROM t_fournisseur WHERE ID_fournisseur = %(value_id_film)s"""
-            with DBconnection() as mydb_conn:
-                mydb_conn.execute(str_sql_genres_films_delete, valeur_select_dictionnaire)
-                data_film_delete = mydb_conn.fetchall()
-                print("data_film_delete...", data_film_delete)
-                session['data_film_delete'] = data_film_delete
+                if not data_film_delete:
+                    # Affiche un message d'avertissement si aucun film n'est trouvé avec l'ID donné
+                    flash("Aucun film trouvé avec cet ID.", "warning")
+                    return redirect(url_for("films_afficher", id_film_sel=0))
 
             btn_submit_del = False
 
-    except Exception as Exception_film_delete_wtf:
+    except Exception as Exception_fournisseur_delete_wtf:
+        # Gère les exceptions et les erreurs
         raise ExceptionFilmDeleteWtf(f"fichier : {Path(__file__).name}  ;  "
                                      f"{film_delete_wtf.__name__} ; "
-                                     f"{Exception_film_delete_wtf}")
+                                     f"{Exception_fournisseur_delete_wtf}")
 
+    # Rend le modèle HTML avec les données récupérées et les messages Flash
     return render_template("films/film_delete_wtf.html",
                            form_delete_film=form_delete_film,
                            btn_submit_del=btn_submit_del,
-                           data_film_del=data_film_delete
-                           )
+                           data_film_del=data_film_delete)
